feat(shares): Phase 224 FTP 共享删除 API (DELETE /api/v1/shares/ftp/{id})
- 使用 SqliteShareRepository 真实数据库实现
- JWT 认证，仅 admin 用户可访问
- 验证共享 ID 存在性（404 Not Found）
- 验证协议为 FTP（非 FTP 返回 404）
- 删除成功返回 204 No Content
- 添加 docs/shares_ftp_delete_api.md 文档

+++ src/handlers/shares_ftp_delete.rs
@@ -89,7 +89,7 @@ pub async fn delete_ftp_share(
     };
 
     // 5. 使用数据库删除共享
-    match repo.delete_share(share_id as i64) {
+    match repo.delete_share(share_id) {
         Ok(true) => {
             // 6. 返回删除成功（204 No Content）
             Ok(HttpResponse::NoContent().finish())
